enyard-admin/backend/requirements.txt

fastapi==0.110.0
uvicorn
mysql-connector-python
pydantic<2.8
